Moved stuff to folders
(cherry picked from commit ee7ec531244efe6a6005cadca389e6132cc5d8b2)
